Create new .h files to organize the algorithm. Program compiling - Finally
