+++ Makefile
@@ -1,4 +1,4 @@
-OBJECTS = limits.o types.o energy.o main.o
+OBJECTS = limits.o readatomslinks.o energy.o spin.o atoms.o links.o main.o
 
 main: ${OBJECTS}
 	g++ -o main ${OBJECTS}
@@ -6,11 +6,20 @@ main: ${OBJECTS}
 limits.o: limits.cpp limits.h
 	g++ -std=c++11 -c limits.cpp
 
-types.o: types.cpp types.h
-	g++ -std=c++11 -c types.cpp
+readatomslinks.o: readatomslinks.cpp readatomslinks.h
+	g++ -std=c++11 -c readatomslinks.cpp
 
 energy.o: energy.cpp energy.h
 	g++ -std=c++11 -c energy.cpp
+
+spin.o: spin.cpp spin.h
+	g++ -std=c++11 -c spin.cpp
+
+atoms.o: atoms.cpp atoms.h
+	g++ -std=c++11 -c atoms.cpp
+
+links.o: links.cpp links.h
+	g++ -std=c++11 -c links.cpp
 
 main.o: main.cpp utilities.h
 	g++ -std=c++11 -c main.cpp
