+++ atoms.cpp
@@ -1,0 +1,9 @@
+#include <iostream>
+
+#include "atoms.h"
+
+std::ostream& operator<< (std::ostream& os, Atom& a)
+{
+    os << "Atom: " << a.id << " of type: " << a.type;
+    return os;
+}
